\name{ncp.pool.pred}
\alias{ncp.pool.pred}
\docType{data}
\title{
prediction model of the non-centrality parameter from simulation results}
\description{
the prediction model, summarized from the simulation results}
\usage{data(ncp.pool.pred)}
\format{
  The format is:
List of 2
 $ : num [1:4] 0.02 0.05 0.1 1
 $ :List of 4
  ..$ : Named num [1:29] -1.067976 19.704888 1.217199 -2.612229 -0.000886 ...
  .. ..- attr(*, "names")= chr [1:29] "(Intercept)" "MAF" "log(OR)" "error" ...
  ..$ : Named num [1:29] -1.11889 4.47646 1.57193 -2.83448 -0.00108 ...
  .. ..- attr(*, "names")= chr [1:29] "(Intercept)" "MAF" "log(OR)" "error" ...
  ..$ : Named num [1:29] -1.152848 2.57742 1.816014 -1.521528 -0.000948 ...
  .. ..- attr(*, "names")= chr [1:29] "(Intercept)" "MAF" "log(OR)" "error" ...
  ..$ : Named num [1:29] -1.17164 0.65682 1.89042 -0.81605 -0.00127 ...
  .. ..- attr(*, "names")= chr [1:29] "(Intercept)" "MAF" "log(OR)" "error" ...
}
\examples{
data(ncp.pool.pred)
## maybe str(ncp.pool.pred) ; plot(ncp.pool.pred) ...
}
\keyword{datasets}
